+++ src/commands/actions/index.ts
@@ -4,6 +4,8 @@ import { action as restartServer } from './restart-server';
 import { action as showOnboardPanel } from './show-onboard-panel';
 import { action as initConfigFile } from './init-config-file';
 import { action as openSettings } from './open-settings';
+import { action as refreshActiveFilePanel } from './refresh-active-file-panel';
+import { action as refreshProjectPanel } from './refresh-project-panel';
 
 export {
     restartServer,
@@ -12,4 +14,6 @@ export {
     stopServer,
     initConfigFile,
     openSettings,
+    refreshProjectPanel,
+    refreshActiveFilePanel,
 };